src/test/java/com/webuitest/suite/DisneySignupTest.java
package com.webuitest.suite;

import io.qameta.allure.Feature;
import org.apache.commons.lang.RandomStringUtils;
import org.openqa.selenium.By;
import org.openqa.selenium.NoSuchElementException;
import org.openqa.selenium.WebDriver;
import org.openqa.selenium.WebElement;
import org.openqa.selenium.support.PageFactory;
import org.openqa.selenium.support.ui.ExpectedConditions;
import org.openqa.selenium.support.ui.FluentWait;
import org.openqa.selenium.support.ui.Wait;
import org.openqa.selenium.support.ui.WebDriverWait;
import org.testng.annotations.Listeners;
import org.testng.annotations.Test;
import pages.DisneyLogin;
import pages.DisneySignup;

import java.io.IOException;
import java.time.Duration;
import java.util.function.Function;

@Feature("Signup and Login Page")

@Listeners({com.listener.Listener.class})

public class DisneySignupTest extends BaseTest {

    // randomize the email generation
    String email = RandomStringUtils.randomAlphanumeric(8) + "@yopmail.com";

    public static By WISHLIST = By.xpath("/descendant::*[contains(@class, 'wishlist-header')][3]");
    public By EMAIL_VERIFY = By.xpath("//*[@class='account-email']");
    public By ACCOUNT_NAME = By.xpath("//*[@class='account-name']");

    @Test(priority = 1)
    /*
     @Stories("Signup Disney")
     @Description("Navigate to Disney page,signup and store signup data in csv file")
     @Step("case1")
     Initialize the page Objects using PageFactory
     Call the methods from the page objects to execute the test
      */

    public void Signup() throws InterruptedException {
        DisneySignup initElements = PageFactory.initElements(browser, DisneySignup.class);
        initElements.clickShop();
        //    initElements.clickHamburger();
        initElements.clickSignup();
        initElements.clickCreateAccount();
        initElements.signupFname("Naveen");
        initElements.signuplname("Krishnan");
        initElements.sigupemail(email);
        initElements.sigupPaswd("Test1234!");
        initElements.sigupVerifyPswd("Test1234!");
        initElements.sigupDoB("01/01/1990");
        initElements.createAcct();
        initElements.continueBtn();
        initElements.signout();

        // write the data to csv file
        initElements.writecsvFile("Naveen", "Krishnan", email);


    }

    @Test(priority = 2)
    public void Login() throws InterruptedException, IOException {
        //  String baseUrl1 = "https://www.shopdisney.com/secure/account";
        DisneySignup signup = PageFactory.initElements(browser, DisneySignup.class);
        // signup.clickShop();  //use this if you are running this test by itself and not from the top
        signup.clickSignup();

        DisneyLogin login = PageFactory.initElements(browser, DisneyLogin.class);
        String email = login.readcsvFile();    // read the csv file for the data stored from previous test case
        System.out.println("from file " + email);
        login.loginEmail();     // populate the email from the data read from excel file
        login.loginPasswd();
        login.loginButton();

        // show casing how a element can be found and appropriate message can be displayed
        boolean element_present = login.verifyLogin();
        if (element_present == true) {
            System.out.println("Login Successful");

        } else {
            System.out.println("Problem with Login");
        }


        // fluent wait just to verify if wishlist element on the header section is located
        Wait<WebDriver> wait = new FluentWait<WebDriver>(driver)
                .withTimeout(Duration.ofSeconds(30))
                .pollingEvery(Duration.ofSeconds(5))
                .ignoring(NoSuchElementException.class);


        WebElement foo = wait.until(new Function<WebDriver, WebElement>() {
            public WebElement apply(WebDriver driver) {
                return driver.findElement(WISHLIST);

            }

        });
        Thread.sleep(20000);
        login.myAccount();
        Thread.sleep(10000);


        driver.switchTo().defaultContent();
        final Wait<WebDriver> wait1 =
                        new WebDriverWait(driver, 30);
                wait1.until(ExpectedConditions.visibilityOfElementLocated(EMAIL_VERIFY));
        Thread.sleep(5000);
        WebElement assert_email = driver.findElement(EMAIL_VERIFY);
        String email_verify = assert_email.getText();
        System.out.println("Email Verified " + email_verify);

        // The below expected result is scraping also the driver. This could be due to switching of driver.. Need to investigate.
      //  Assert.assertEquals(assert_email,  email_verify);   // asserting the email on the page and from csv file



    }

}